Add automated browser testing
This adds a `test-saucelabs` script to package.json that uses [zuul] to
run tests on browsers using [SauceLabs].

Zuul is setup to test the browsers in the current [browserslist] defaults
list (shown at <http://browserl.ist>), constrained to what zuul
supports, according to:

    ./node_modules/.bin/zuul --list-available-browsers

(Opera 11 fails on SauceLabs, so skip it)

---

Travis CI is configured to run these tests if the `SAUCE_USERNAME` and
`SAUCE_ACCESS_KEY` environment variables are set. These credentials can
be obtained for free by [signing up] for the SauceLabs Open Source
program. Travis offers a way to set environment variables in the [build
settings], or they can be [encrypted] and added to the .travis.yml file.

Here's what a build looks like with this change:
https://travis-ci.org/josephfrazier/xregexp/builds/220784295

---

This also includes a slight refactor that moves `beforeEach` calls from
test helpers into spec files, because having them in the helpers doesn't
work with zuul. There is a `test-browser` script that can be used to run
zuul locally.

[zuul]: https://github.com/defunctzombie/zuul
[SauceLabs]: https://saucelabs.com/open-source
[browserslist]: https://github.com/ai/browserslist
[signing up]: https://saucelabs.com/beta/signup/OSS/None
[build settings]: https://travis-ci.org/slevithan/xregexp/settings
[encrypted]: https://docs.travis-ci.com/user/environment-variables/#Defining-encrypted-variables-in-.travis.yml